fix delete and crud property category

+++ src/Public/Components/AdminApp/PromotionDetail/PromotionDetailList/PromotionDetailList.js
@@ -58,6 +58,10 @@ function PromotionDetailList() {
     };
 
     const handleDelete = async (event, id) => {
+        if (!window.confirm('Are you want to delete?')){
+            return;
+        }
+
         event.preventDefault();
 
         await promotionDetailService.delete(id)
